NLP AND COSINE SIMILARITY
In this notebook the code is trying to find duplicate of a given sentence. We have a excel which has been prepareped by scraping a internal website containing 1000 of ideas and description/details of the ideas. We need to find if a given idea is a duplicate of any given ideas