Working with excel, selecting dates sites and performing calculation
In this RPA I am extracting data from multiple excel based on some specific inputs like data for specific dates, the site for which data is required, max. or min. data volume reqd. etc and pasting into my final calculation. With the help of this BOT the time required to perform the activity was reduced to 15 mins from 2hrs as it has reduced a tons of calcuation involved.